# tacocloud/tacocloud-api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Taco Cloud Public API
  version: 1.0.0
  description: |
    Contrato oficial OpenAPI 3.0 para Taco Cloud (Laboratorio 6 - TC-35).
    Define los contratos de integración para ingredientes, diseño de tacos, pedidos con idempotencia
    y trazabilidad por Correlation ID, favoritos, cupones, pasarela de pago tokenizada, operaciones de cocina y
    anuncios operativos seguros. Incluye el modelo canónico de error RFC 7807 (ApiProblem).
  contact:
    name: Taco Cloud Engineering Team
    email: support@tacocloud.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desarrollo local (v1 Oficial)
  - url: http://localhost:8080/api
    description: Servidor de desarrollo local (Alias legado - Deprecado)
...
paths:
...
  /tacos:
    get:
      tags: [Tacos]
      summary: Buscar y filtrar tacos con paginación
      operationId: searchTacos
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: diet
          in: query
          required: false
          schema:
            type: string
            enum: [VEGAN, VEGETARIAN, GLUTEN_FREE, KETO]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Página de tacos encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TacoPageResponse'
    post:
      tags: [Tacos]
      summary: Diseñar y guardar un nuevo taco
      operationId: postTaco
      security:
        - basicAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TacoRequest'
      responses:
        '201':
          description: Taco diseñado y guardado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TacoResponse'
        '400':
          description: Violación de reglas de diseño o campos requeridos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
...
  /tacos/top:
    get:
      tags: [Tacos]
      summary: Consultar ranking de tacos más votados
      operationId: getTopTacos
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista de tacos mejor puntuados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopTacoResponse'
  /tacos/validate:
    post:
      tags: [Tacos]
      summary: Validar viabilidad física de un diseño de taco
      operationId: validateTacoDesign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TacoDesignRequest'
      responses:
        '200':
          description: Resultado de la validación física
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TacoDesignValidationResponse'
  /tacos/{id}:
    get:
      tags: [Tacos]
      summary: Consultar detalle de un taco por ID
      operationId: getTacoById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Taco encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TacoResponse'
        '404':
          description: Taco no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
  /tacos/{id}/rating:
    get:
      tags: [Tacos]
      summary: Consultar promedio y votos de un taco
      operationId: getTacoRating
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resumen de valoraciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TacoRatingSummaryResponse'
    put:
      tags: [Tacos]
      summary: Emitir o actualizar puntuación para un taco (1 a 5)
      operationId: rateTaco
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingRequest'
      responses:
        '200':
          description: Calificación registrada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TacoRatingSummaryResponse'
        '400':
          description: Puntuación inválida (debe ser entre 1 y 5)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para usuarios autenticados
    basicAuth:
      type: http
      scheme: basic
      description: Autenticación básica HTTP (usuario y contraseña)
...
  schemas:
...
    IngredientResponse:
      type: object
      required:
        - id
        - name
        - type
      properties:
        id:
          type: string
          example: FLTO
        name:
          type: string
          example: Flour Tortilla
        type:
          type: string
          example: WRAP

    # --- TACOS ---
    TacoRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 5
          example: Carnitas Fiesta
        ingredientIds:
          type: array
          items:
            type: string
          example: [FLTO, CARN, CHED]
    TacoResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          example: 64b8a1c9e4b0c23a8f123456
        name:
          type: string
          example: Carnitas Fiesta
        createdAt:
          type: string
          format: date-time
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientResponse'
        dietaryTags:
          type: array
          items:
            type: string
          example: [GLUTEN_FREE]
        allergens:
          type: array
          items:
            type: string
          example: [DAIRY]
        spiceLevel:
          type: string
          example: MEDIUM
    TacoPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TacoResponse'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 45
        totalPages:
          type: integer
          example: 3
    TacoDesignRequest:
      type: object
      required:
        - name
        - ingredientIds
      properties:
        name:
          type: string
          example: Supremo
        ingredientIds:
          type: array
          items:
            type: string
          example: [FLTO, GRBF, CHED, SLSA]
    TacoDesignValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
          example: true
        violations:
          type: array
          items:
            type: string
          example: []
    RatingRequest:
      type: object
      required:
        - score
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
    TacoRatingSummaryResponse:
      type: object
      properties:
        tacoId:
          type: string
          example: 64b8a1c9e4b0c23a8f123456
        averageScore:
          type: number
          format: double
          example: 4.8
        totalVotes:
          type: integer
          example: 120
        userScore:
          type: integer
          example: 5
    TopTacoResponse:
      type: object
      properties:
        tacoId:
          type: string
          example: 64b8a1c9e4b0c23a8f123456
        tacoName:
          type: string
          example: Carnitas Fiesta
        averageScore:
          type: number
          example: 4.9
        totalVotes:
          type: integer
          example: 350

    # --- ÓRDENES ---
...
    ApiProblem:
      type: object
      required:
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Request validation failed
        instance:
          type: string
          example: /api/v1/orders
        code:
          type: string
          example: VALIDATION_ERROR
        violations:
          type: array
          items:
            $ref: '#/components/schemas/ApiProblemViolation'
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
    ApiProblemViolation:
      type: object
      properties:
        field:
          type: string
          example: deliveryName
        message:
          type: string
          example: Delivery name is required